refactor: unify `CoreCryptoContext` mls impls
# Conflicts:
#	crypto-ffi/src/generic/context/e2ei.rs
#	crypto-ffi/src/generic/context/mod.rs
#	crypto-ffi/src/generic/context/proteus.rs
#	crypto-ffi/src/generic/mod.rs
