+++ crypto-ffi/src/wasm/mod.rs
@@ -1,3 +1,2 @@
 #![allow(unused_variables)]
 pub mod context;
-mod utils;
